create `api` folder in components; move over and rename api components
